% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_var_names.R
\name{apply_var_names}
\alias{apply_var_names}
\title{(re)name climate variable}
\usage{
apply_var_names(climate, var_names)
}
\arguments{
\item{climate}{an object of class tc_climate}

\item{var_names}{either NULL or a character vector}
}
\value{
the (potentially renamed) climate data
}
\description{
(re)name climate variable
}
\keyword{internal}
