\name{.range}
\alias{.mean}
\alias{.range}
\alias{.sum}
\title{Modifiers for climate parameter selection}
\usage{
.range(.months = NULL, .variables = NULL)

.mean(.months = NULL, .variables = NULL)

.sum(.months = NULL, .variables = NULL)
}
\arguments{
  \item{.months}{}

  \item{.variables}{}
}
\description{
These modifiers are used to select specific months from
specific climate parameters, and potentially transform the
selections into their respective sums or means. The
modifiers can be chained together using `+`.
}

