% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pmat.R
\name{make_pmat}
\alias{make_pmat}
\title{Make list of parameter matrices from climate input data}
\usage{
make_pmat(x, pad = 0)
}
\arguments{
\item{x}{climate data as returned from as_tcclimate}

\item{pad}{how many previous years should be padded with NAs? This is useful when
we want to use the full range of available data when all predictors derive
from the current year (pad = 2, or current and previous, for pad = 1)}
}
\value{
a list of matrices
}
\description{
Reformat climate input data into separate lists
}
\details{
Make separate lists for each parameter in the climate data input, and make
them accessible by the parameter names. The single lists correspond to the
scheme years in rows, observations from previous december to current
december in columns.
}
\keyword{internal}
\keyword{manip}

