\docType{data}
\name{muc.clim}
\alias{muc.clim}
\title{Monthly Mean Temperature and Total Precipitation for Forstenrieder Park,
Munich}
\format{A \code{data.frame} containing four columns with year, month,
  temperature and precipitation.}
\source{
Zang C, Pretzsch H, Rothe A (2012) Size-dependent responses
to summer drought in Scots pine, Norway spruce and common
oak. Trees - Structure and Function, 26, 557-569.
}
\usage{
data(muc.clim)
}
\description{
This dataset gives the monthly mean temperature and total
precipitation at Forstenrieder Park, Munich, Bavaria,
Germany.
}
\keyword{datasets}

