% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{muc_clim}
\alias{muc_clim}
\title{Monthly Mean Temperature and Total Precipitation for Forstenrieder
Park, Munich}
\format{A \code{data.frame} containing four columns with year,
month, temperature and precipitation.}
\source{
Zang C, Pretzsch H, Rothe A (2012) Size-dependent
responses to summer drought in Scots pine, Norway spruce and
common oak. Trees - Structure and Function, 26, 557-569.
}
\usage{
data(muc_clim)
}
\description{
This dataset gives the monthly mean temperature and total
precipitation at Forstenrieder Park, Munich, Bavaria, Germany.
}
\keyword{datasets}

