% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norw015}
\alias{norw015}
\title{Tree-Ring Chronology of a Pine Population at Visdalen, Norway}
\format{A \code{data.frame} containing tree-ring indices and
replication depth with respective years as \code{rownames}.}
\source{
Briffa K. Briffa - Visdalen - PISY - ITRDB
NORW015. http://hurricane.ncdc.noaa.gov/pls/paleox/f?p=519:1:::::P1_STUDY_ID:2861. Accessed
2014/07/03.
}
\usage{
data(norw015)
}
\description{
This dataset gives the tree-ring indices for \emph{Pinus
sylvestris} at a site near Visdalen, Norway. The chronology
represents 34 dated series and spans the period between 1600 and
1983. The raw measurements were obtained from the ITRDB (see
source), and read in, detrended with cubic splines (frequency
cutoff of 50 percent at two thirds of the series length) and
averaged using \code{\link[dplR]{read.rwl}},
\code{\link[dplR]{detrend}}, and \code{\link[dplR]{chron}} from
package \code{dplR}.
}
\keyword{datasets}

