\docType{data}
\name{rt.temp}
\alias{rt.temp}
\title{Monthly Temperature Means for Rothenburg ob der Tauber, Germany}
\format{A \code{data.frame} containing thirteen columns with year
and twelve months of temperature data in degree Celsius.}
\source{
Zang C, Rothe A, Weis W, Pretzsch H (2011) Zur
Baumarteneignung bei Klimawandel: Ableitung der
Trockenstress-Anfälligkeit wichtiger Waldbaumarten aus
Jahrringbreiten. Allgemeine Forst- und Jagdzeitung, 182,
98-112.
}
\usage{
data(rt.temp)
}
\description{
This dataset gives the monthly temperature means at
Rothenburg ob der Tauber, Bavaria, Germany in a decadal
(DENDROCLIM2002-style) format)
}
\keyword{datasets}

