% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rt_prec}
\alias{rt_prec}
\title{Monthly Precipitation Sums for Rothenburg ob der Tauber, Germany}
\format{
A \code{data.frame} containing thirteen columns with year
and twelve months of precipitation data in mm rainfall.
}
\source{
Zang C, Rothe A, Weis W, Pretzsch H (2011) Zur
Baumarteneignung bei Klimawandel: Ableitung der
Trockenstress-Anfaelligkeit wichtiger Waldbaumarten aus
Jahrringbreiten. Allgemeine Forst- und Jagdzeitung, 182, 98-112.
}
\usage{
data(rt_prec)
}
\description{
This dataset gives the monthly precipitation sum at Rothenburg ob der Tauber,
Bavaria, Germany in a decadal (DENDROCLIM2002-style) format)
}
\keyword{datasets}
