% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_year.R
\name{shift_year}
\alias{shift_year}
\title{shifts a month 0, 1, or 2 years back}
\usage{
shift_year(month, by = 0)
}
\description{
there is no intuitive way to signify months from the year previous to the
last one...
}
\keyword{internal}

