% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{spai020}
\alias{spai020}
\title{Tree-Ring Chronology of a Pine Population at Penota, Spain}
\format{A \code{data.frame} containing tree-ring indices and
replication depth with respective years as \code{rownames}.}
\source{
Fernandez-Cancio A, Genova Fuster M: Fernandez-Cancio -
Penota - PISY - ITRDB SPAI020,
http://hurricane.ncdc.noaa.gov/pls/paleox/f?p=519:1:::::P1_STUDY_ID:3255,
accessed 2014/07/03.
}
\usage{
data(spai020)
}
\description{
This dataset gives the tree-ring indices for \emph{Pinus
sylvestris} at a site near Penota, Spain. The chronology
represents 18 dated series and spans the period between 1763 and
1991. The chronology was obtained from the ITRDB (see source), and
read in using \code{\link[dplR]{read.crn}} from package
\code{dplR}.
}
\keyword{datasets}

