% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spain_prec}
\alias{spain_prec}
\title{Monthly Precipitation Sums for Spain}
\format{
A \code{data.frame} containing thirteen columns with year
and twelve months of temperature data in degree Celsius.
}
\source{
Mitchell TD, Carter TR, Jones PD, Hulme M, New M, et
al. (2004) A comprehensive set of high-resolution grids of monthly
climate for Europe and the globe: the observed record (1901-2000)
and 16 scenarios (2001-2100). Tyndall Centre for Climate Change
Research Working Paper 55, 25.
}
\usage{
data(spain_prec)
}
\description{
This data set gives the monthly precipitation sums at country
level for Spain in a decadal (DENDROCLIM2002-style) format from
the TYN CY 1.1 data set.
}
\keyword{datasets}
