% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{spain_temp}
\alias{spain_temp}
\title{Monthly Temperature Means for Spain}
\format{A \code{data.frame} containing thirteen columns with year
and twelve months of temperature data in degree Celsius.}
\source{
Mitchell TD, Carter TR, Jones PD, Hulme M, New M, et
al. (2004) A comprehensive set of high-resolution grids of monthly
climate for Europe and the globe: the observed record (1901-2000)
and 16 scenarios (2001-2100). Tyndall Centre for Climate Change
Research Working Paper 55, 25.
}
\usage{
data(spain_temp)
}
\description{
This data set gives the monthly temperature means at country level
for Spain in a decadal (DENDROCLIM2002-style) format from the TYN
CY 1.1 data set.
}
\keyword{datasets}

