% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tc_correlation}
\alias{tc_correlation}
\title{(bootstrapped) correlation function}
\usage{
tc_correlation(chrono, climate, ci, boot)
}
\arguments{
\item{chrono}{a tree-ring chronology}

\item{climate}{data.frame with climate parameters}

\item{ci}{numeric: p-level for confidence interval (must be in
c(0.1, 0.05, 0.01)}

\item{boot}{character, which bootstrapping method should be used
(or "none")}

\item{p}{probability for rgeom, that determines distribution of
sampling blocks for stationary bootstrap scheme}
}
\description{
See documentation of dcc for details.
}
\keyword{internal}

