% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_design.R
\name{tc_design}
\alias{tc_design}
\title{Deparse list structure from month specification and return parameter set}
\usage{
tc_design(selection, climate, check_2 = TRUE)
}
\arguments{
\item{selection}{the list structure to specify the month selection}

\item{climate}{the climate data as returned by make_pmat}

\item{check_2}{should there be a check for less than 2 variables
in the design matrix?}
}
\value{
a data.frame
}
\description{
Deparse list structure from month specification into single calls, call an
aggregating function to collect the data, and return a parameter set for
calibration. In general, the months must be specified as a list or nested
list structure. Possible cases are:
list()
list(...)
list(list(...))
list(list(...), list(...), ...)
}

