% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_selection.R
\name{unify_selection}
\alias{unify_selection}
\title{unify selections to handle "classic" (bootRes-style) selections}
\usage{
unify_selection(selection)
}
\arguments{
\item{selection}{a treeclim selection}
}
\value{
unified (list-style) selection
}
\description{
unify selections to handle "classic" (bootRes-style) selections
}
\keyword{internal}
