% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ac}
\alias{ac}
\title{Report autocorrelation structure}
\usage{
ac(x)
}
\arguments{
\item{x}{an object of class 'tc_dcc'}
}
\value{
nothing, invoked for side effects (reporting
autocorrelation structure)
}
\description{
Print the autocorrelation structure of the original proxy data
series and its bootstrapped samples.
}

