% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tcclimate.R
\name{as_tcclimate}
\alias{as_tcclimate}
\title{Create the correct data format for variable selection}
\usage{
as_tcclimate(x, varnames = NULL)
}
\arguments{
\item{x}{the supplied climate data}

\item{varnames}{optionally supplied variable names}
}
\value{
a data.frame with years in rows and monthly observations
in columns
}
\description{
Check the climate data that is either supplied as a matrix, a
data.frame, or a list of data.frames. Reformat the data for
further usage into the internal format.
}
\keyword{internal}
\keyword{manip,}
