% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_months.R
\name{check_months}
\alias{check_months}
\title{Check if all months are between -24 and 12 (previous and current december)}
\usage{
check_months(selection)
}
\arguments{
\item{selection}{the month selection as vector, list, or nested
list}
}
\value{
a list with the earliest month and the result of the check
(logical)
}
\description{
Make sure that all the months specified in the (probably complex)
month specification to dcc are correct months, in the sense that
they range between -24 and 12. Further, return the earliest month
in the complete selection to decide if we need previous year
climate later on.
}
\keyword{internal}

